requirement.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt